pyproject.toml
[project]
name = "sqlalchemy-encryption"
version = "1.0"
dependencies = [
    "alchemical",
    "cryptography",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
